% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fMCMC_Net.R
\name{mcmcDirNet}
\alias{mcmcDirNet}
\title{Simulate the posterior distribution of the parameters of an exponential random symmetric graph model}
\usage{
mcmcDirNet(
  network,
  formula.u,
  formula.v,
  formula.w,
  f_to_var.u,
  f_to_var.v,
  f_to_var.w,
  heterogeneity,
  starting = NULL,
  init.network = NULL,
  nblock = 2,
  ncores = 1,
  prior = list(),
  Sigmatheta = NULL,
  Sigmahete = NULL,
  simutheta = 1000,
  simunet = 1000,
  mcmc.ctr = list(target = NULL, kappa = 0.6, jmin = 1e-06, jmax = 3),
  data = NULL
)
}
\description{
Simulate the posterior distribution of the parameters of an exponential random symmetric graph model
}
