% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcERGM.R
\docType{package}
\name{mcmcERGM-package}
\alias{mcmcERGM}
\alias{mcmcERGM-package}
\title{The mcmcERGM package}
\description{
The \pkg{mcmcERGM} package simulates the posterior distribution of the parameters of an Exponential Random Graph Models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ahoundetoungan/mcmcERGM}
  \item Report bugs at \url{https://github.com/ahoundetoungan/mcmcERGM/issues}
}

}
\author{
\strong{Maintainer}: Elysée Aristide Houndetoungan \email{ariel92and@gmail.com}

Authors:
\itemize{
  \item Yves Zenou \email{yves.zenou@monash.edu}
  \item Michael Vlassopoulos \email{M.Vlassopoulos@soton.ac.uk}
}

}
