% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fMCMC_Net.R
\name{mcmcSymNet}
\alias{mcmcSymNet}
\title{Simulate the posterior distribution of the parameters of an exponential random symmetric graph model}
\usage{
mcmcSymNet(
  network,
  formula.u,
  formula.w,
  f_to_var.u,
  f_to_var.w,
  starting = NULL,
  nblock = 2,
  ncores = 1,
  Etheta = NULL,
  invVtheta = NULL,
  Sigma = NULL,
  simutheta = 1000,
  simunet = 1000,
  mcmc.ctr = list(target = NULL, kappa = 0.6, jmin = 1e-06, jmax = 3),
  data = NULL
)
}
\description{
Simulate the posterior distribution of the parameters of an exponential random symmetric graph model
}
