% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_Net.R
\name{simDirNet}
\alias{simDirNet}
\title{Simulate networks using an exponential random directed graph model}
\usage{
simDirNet(
  formula.u,
  formula.v,
  formula.w,
  f_to_var.u,
  f_to_var.v,
  f_to_var.w,
  theta,
  heterogeneity,
  size = NULL,
  init.net = NULL,
  nblock = 2,
  ncores = 1,
  simunet = 1000,
  fgraph = NULL,
  fsubgraph = NULL,
  data = NULL
)
}
\description{
Simulate networks using an exponential random directed graph model
}
